Extract utilities into 'core' package
